% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augere.core-package.R
\docType{package}
\name{augere.core-package}
\alias{augere.core}
\alias{augere.core-package}
\title{augere.core: Core Utilities for Automatically Generated Reports}
\description{
Provides utility functions for automatic generation of reports in the augere framework. This mostly involves parsing and dynamically editing Rmarkdown files, controlling inputs into the augere pipelines. Each pipeline generates and executes a self-contained Rmarkdown file for a common bioinformatics analysis; see downstream packages like augere.de and augere.screen for examples of some analysis pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/augere-bioinfo/augere.core}
  \item Report bugs at \url{https://github.com/augere-bioinfo/augere.core/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})

}
