% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparseToString.R
\name{deparseToString}
\alias{deparseToString}
\title{Deparse object to a string}
\usage{
deparseToString(x, indent = NULL)
}
\arguments{
\item{x}{R object to be deparsed.}

\item{indent}{Integer scalar specifying the indenting for multi-line deparsed output.
If \code{NULL}, the deparsed output will be put on a single line with no indenting.}
}
\value{
String containing the deparsed value of \code{x}.
}
\description{
Deparse an object to a single string by concatenation, possibly with indentation for multi-line output.
}
\details{
This function is based on \code{\link{deparse}} but with special accommodations when the deparsed output spans multiple lines.
If \code{indent=NULL}, the multiple output lines are directly \code{paste}d together with no separator. 
If \code{indent} is an integer, a newline is added between lines, indented by the specified number of spaces.
}
\examples{
x <- sample(100)
deparse(x)
deparseToString(x)
deparseToString(x, indent=4)

}
\author{
Aaron Lun
}
