% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateChunks.R
\name{evaluateChunks}
\alias{evaluateChunks}
\title{Evaluate R chunks}
\usage{
evaluateChunks(chunks, env)
}
\arguments{
\item{chunks}{List of character vectors, typically from \code{\link{extractChunks}}.
Each vector corresponds to a chunk of R code.}

\item{env}{Environment in which to evaluate \code{chunks}.}
}
\value{
All \code{chunks} are evaluated within \code{env}.
}
\description{
Evaluate R chunks, typically derived from an Rmarkdown report.
}
\details{
Each character vector may have some \pkg{knitr}-inspired attributes:
\itemize{
\item \code{eval} is a string that, upon evaluation within \code{env}, specifies if the current chunk should be evaluated.
\item \code{ref.label} is a string that, upon evaluation within \code{env}, contains the name of another character vector in \code{chunks}.
The chunk in the named character vector will be evaluated instead of the current chunk.
\item \code{error} is a string that, upon evaluation within \code{env}, specifies if errors should be ignored for this chunk.
}
}
\examples{
chunks <- list("a <- 1", "a <- a + 2", "", "a <- NULL")
names(chunks) <- c("", "foo1", "", "foo3")
attr(chunks[[3]], "ref.label") <- "paste0('foo', 1)"
attr(chunks[[4]], "eval") <- "a < 0"

env <- new.env()
evaluateChunks(chunks, env)
env$a

}
\author{
Aaron Lun
}
