% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractChunks.R
\name{extractChunks}
\alias{extractChunks}
\title{Extract R chunks}
\usage{
extractChunks(contents)
}
\arguments{
\item{contents}{A character vector or list (of lists) of character vectors containing an Rmarkdown document,
typically obtained via \code{\link{parseRmdTemplate}}.
Each string should correspond to a separate line.}
}
\value{
List of character vectors, each of which contains the R code for a single chunk.
Each vector is named according to the name of the chunk.
\pkg{knitr} properties are stored in the attributes for each vector.
}
\description{
Extract R chunks from an Rmarkdown document, along with their names and \pkg{knitr} properties.
}
\examples{
contents <- "
```{r alpha}
alpha <- 1
bravo <- 2
```

Some random text here

```{r, echo=FALSE, eval=FALSE}
charlie <- 3
delta <- 4
```

- point 1 with some `r inline <- TRUE`.
- point 2.

  ```{r i-am-indented, error=TRUE}
  echo <- 5
  foxtrot <- 6
  ```

even more text here with `r 'some'` `r 'more'` `r 'inlining'`.
"

extractChunks(strsplit(contents, "\n"))

}
\author{
Aaron Lun
}
