% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRmdTemplate.R
\name{parseRmdTemplate}
\alias{parseRmdTemplate}
\title{Parse Rmarkdown template}
\usage{
parseRmdTemplate(contents)
}
\arguments{
\item{contents}{String or character vector containing an Rmarkdown report template.}
}
\value{
List of character vectors or nested lists, see Details.
}
\description{
Parse an Rmarkdown report template to extract blocks of text or code.
}
\details{
An Rmarkdown report template may contain any number of paired \code{:BEGIN} and \code{:END} tags.
These tags enclose blocks of text that may be dynamically removed or duplicated by the pipeline functions.

Tag pairs should follow these rules:
\itemize{
\item Each tag should start on its own line.
\item The \code{:BEGIN} tag should be followed by a name, e.g., \code{:BEGIN my-name-here}.
\item The corresponding \code{:END} tag does not need to be followed by a name, but if it does, the name should be the same as that of the matching \code{:BEGIN} tag.
\item Tag pairs may be nested within another tag pair.
\item Names should be unique for any given level of nesting.
For nested tag pairs, there should be no other tag pair with the same name within the parent tag pair.
For unnested tag pairs, there should be no other unnested tag pair with the same name. 
}

In the output of this function, each entry of the top-level list may be either:
\itemize{
\item A list, corresponding to a tag pair.
This list contains all text in the Rmarkdown template enclosed by the tag pair,
which may be character vectors or further lists corresponding to nested tag pairs.
If the tag pair is named, the same name will be used for this list entry. 
\item A character vector, to store newline-separated text within or between tag pairs.
} 
Calling \code{unlist} on the output of this function will return a newline-separated version of \code{contents} without the tags.
}
\examples{
template <- "
Ochite iku sunadokei bakari miteru yo
Sakasama ni sureba hora mata hajimaru yo
Kizanda dake susumu jikan ni
Itsuka boku mo haireru kana

```{r}
nagisa <- 'furukawa' 
```

:BEGIN clannad

Kimi dake ga sugisatta saka no tochuu wa
Atataka na hidamari ga ikutsu mo dekiteta
Boku hitori ga koko de yasashii
Atatakasa wo omoikaeshiteru

:BEGIN after-story

```{r}
tomoya <- 'Okazaki'
```

:END

Kimi dake wo kimi dake wo
Suki de ita yo
Kaze de me ga nijinde
Tooku naru yo

:BEGIN toki-wa-kizamu-uta

Itsumademo oboeteru
Kono machi ga kawatte mo
Dore dake no kanashimi to deau koto ni natte mo
Misete yaru hontou wa tsuyokatta doki no koto
Saa iku yo arukidasu saka no michi wo

```{r}
ushio <- TRUE
```

:END

:END"

parseRmdTemplate(template)
}
\author{
Aaron Lun
}
