% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacePlaceholders.R
\name{replacePlaceholders}
\alias{replacePlaceholders}
\title{Substitute placeholders}
\usage{
replacePlaceholders(contents, replacements)
}
\arguments{
\item{contents}{String, character vector or list containing an Rmarkdown report template.
This is typically the output of \code{\link{parseRmdTemplate}}.}

\item{replacements}{Named character vector of replacement text for placeholders.
The name is the placeholder and the value is the replacement text.
This may also be a named list of strings.}
}
\value{
An object of the same type as \code{contents}, but with all of the placeholders replaced.
}
\description{
Find all placeholders in an Rmarkdown template and replace them with actual text.
}
\details{
A placeholder takes the form \code{<\%= LABEL \%>}, where \code{LABEL} can be any name without \code{\%} or newlines.
Pipeline functions are expected to replace these placeholders with actual text, usually based on user-supplied parameters.
}
\examples{
replacePlaceholders(
    "Hi my name is <\%= NAME \%> and I enjoy <\%= FOOD \%>.",
    replacements=c(NAME="Aaron", FOOD="pork")
)

}
\author{
Aaron Lun
}
