% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResult.R
\name{saveResult}
\alias{saveResult}
\alias{readResult}
\title{Save a DE result}
\usage{
saveResult(x, path, metadata, meta.name = "_metadata.json", requires = NULL)

readResult(path, meta.name = "_metadata.json")
}
\arguments{
\item{x}{Result object to be saved.}

\item{path}{String containing the path to the directory in which \code{x} is to be saved.}

\item{metadata}{Named list of metadata to save alongside \code{x}.
If \code{NULL}, no metadata is saved.}

\item{meta.name}{String containing the name of the metadata file within \code{path}.
Only used if \code{metadata} is not \code{NULL}.}

\item{requires}{String containing the name of the \pkg{alabaster.*} package containing the relevant methods to save \code{x}.
This may be necessary if the class of \code{x} is not known to \code{\link[alabaster.base]{saveObject}}.}
}
\value{
For \code{saveResult}, \code{x} and \code{metadata} (if supplied) are saved to disk at \code{path}.
\code{NULL} is invisibly returned.
}
\description{
Save result objects to disk using the \pkg{alabaster.base} framework.
}
\examples{
df <- S4Vectors::DataFrame(A = 1:5, B=2:6)
meta <- list(
    title = "This is a very important result",
    author = "Myself",
    date = "2022-02-22"
)

tmp <- tempfile()
saveResult(df, tmp, metadata = meta)
list.files(tmp, recursive=TRUE)

readResult(tmp)

}
\seealso{
\code{\link[alabaster.base]{saveObject}} and \code{\link[alabaster.base]{readObject}}, for the underlying functionality.
}
\author{
Aaron Lun
}
