% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRmd.R
\name{writeRmd}
\alias{writeRmd}
\title{Write Rmarkdown file}
\usage{
writeRmd(contents, file, append = FALSE, clean.empty = TRUE)
}
\arguments{
\item{contents}{A character vector of a (nested) list of such vectors.}

\item{file}{String containing a file path or a connection object to an opened file.}

\item{append}{Logical scalar indicating whether to append \code{contents} to \code{file}.}

\item{clean.empty}{Logical scalar indicating whether extraneous empty lines should be removed.}
}
\value{
\code{contents} is written to \code{file} and \code{NULL} is invisibly returned.
}
\description{
Write Rmarkdown-formatted strings to file.
}
\details{
Each element of each character vector in \code{contents} is placed on a new line in the output.
Empty strings are respected and manifest as empty lines.
If \code{clean.empty=TRUE}, repeated empty lines are removed.
}
\examples{
contents <- list(
    c("# TITLE", ""),
    list(
        c("```{r}", "a <- 1", "```"),
        c("", "super foo bar")
    ),
    c("", "## section")
)
tmp <- tempfile(fileext=".Rmd")
writeRmd(contents, tmp)
cat(readLines(tmp), sep="\n") # having a look

}
\author{
Aaron Lun
}
