% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augere.core-package.R
\name{augere.core-package}
\alias{augere.core-package}
\title{augere.core}
\description{
Implement utilities for the augere framework,
to automatically generate reports for common bioinformatics analyses.
See downstream packages like \pkg{augere.de} and \pkg{augere.screen} for available analysis pipelines.
}
